scorelist view, refactoring and styling
