+++ Quicktournament/src/components/PlayerList.tsx
@@ -3,12 +3,9 @@ import React, { useEffect, useState } from "react";
 import { StyleSheet, Text, View } from "react-native";
 import { Button } from "react-native-elements";
 import { useDispatch, useSelector } from "react-redux";
-import { generateWaitingList } from "../../store/reducers/playerSlice";
 import { RootState } from "../../store/store";
 import RealmContext from '../Realm/RealmConfig';
 import { NewPlayer } from "../interfaces/interfaces";
-import { Player } from "../models/Player";
-import { serializeObject } from "../utils/utils";
 import OnGoingGame from "./OnGoingGame";
 import RegisterNewPlayer from "./RegisterPlayer";
 import WaitingList from "./WaitinList";
@@ -25,19 +22,12 @@ const PlayerList = ({ route }) => {
     const [isVisible, setIsVisible] = useState(false);
     const [playersByGameType, setPlayersByGametype] = useState<NewPlayer[]>([]);
 
-    const dispatch = useDispatch();
     const realm = useRealm()
 
     useEffect(() => {
-        //deleteAll();
         setIsVisible(false);
 
-        if(players.length < 1) {
-            const playerList = getPlayers();
-            if(playerList.length) {
-                dispatch(generateWaitingList(playerList));
-            } 
-        } else {
+        if(players.length) {
             const playerList = players.filter((p) => p.gameType === gameType && p.onGoingGame == 0 && p.lost == 0);
             setPlayersByGametype(playerList);
         } 
@@ -51,24 +41,6 @@ const PlayerList = ({ route }) => {
                 <Text>Ladataan pelaajat ja pelit</Text>
             </View>
         );
-    }
-
-    const deleteAll = () => {
-        realm.write(() => {
-            realm.deleteAll();
-        });
-    }
-
-    const getPlayers = () => {
-
-        let playerList: any;
-        playerList = realm.objects<Player>("Player");
-    
-        const players: NewPlayer[] = [];
-        playerList.map((player: NewPlayer) => {
-            players.push(serializeObject(player));
-        });
-        return players;
     }
 
     return (
