+++ Quicktournament/src/screens/Games.tsx
@@ -7,24 +7,44 @@ import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../../store/store";
 import RealmContext from '../Realm/RealmConfig';
 import { generateEndedGamesList } from "../../store/reducers/gameSlice";
+import { NewPlayer } from "../interfaces/interfaces";
+import { generateWaitingList } from "../../store/reducers/playerSlice";
+import { Player } from "../models/Player";
 
 const { useRealm } = RealmContext;
 
 const Games = ({navigation}) => {
 
     const games: Game[] = useSelector((state: RootState) => state.game.games);
+    const players: NewPlayer[] = useSelector((state: RootState) => state.player.players);
 
     const dispatch = useDispatch();
     const realm = useRealm();
 
     useEffect(() => {
+        //deleteAll();
         if(!games.length) {
             const fetchGames = fetchOngoingGame();
             if(fetchGames !== undefined) {
                 dispatch(generateEndedGamesList(fetchGames));
             }
         }
-    })
+    },[]);
+
+    useEffect(() => {
+        if(players.length < 1) {
+            const playerList = getPlayers();
+            if(playerList.length) {
+                dispatch(generateWaitingList(playerList));
+            } 
+        }
+    },[])
+
+    const deleteAll = () => {
+        realm.write(() => {
+            realm.deleteAll();
+        });
+    }
 
     const fetchOngoingGame = () => {
         let allGames: any
@@ -36,6 +56,18 @@ const Games = ({navigation}) => {
         })
 
         return games;
+    }
+
+    const getPlayers = () => {
+
+        let playerList: any;
+        playerList = realm.objects<Player>("Player");
+    
+        const players: NewPlayer[] = [];
+        playerList.map((player: NewPlayer) => {
+            players.push(serializeObject(player));
+        });
+        return players;
     }
 
     return (
